+++ .docker/Dockerfile
@@ -117,7 +117,6 @@ COPY ghost/prometheus-metrics/package.json ghost/prometheus-metrics/package.json
 COPY ghost/recommendations/package.json ghost/recommendations/package.json
 COPY ghost/security/package.json ghost/security/package.json
 COPY ghost/slack-notifications/package.json ghost/slack-notifications/package.json
-COPY ghost/stripe/package.json ghost/stripe/package.json
 COPY ghost/tiers/package.json ghost/tiers/package.json
 COPY ghost/webmentions/package.json ghost/webmentions/package.json
