+++ compose.yml
@@ -85,7 +85,6 @@ services:
       - node_modules_ghost_recommendations:/home/ghost/ghost/recommendations/node_modules:delegated
       - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
       - node_modules_ghost_slack-notifications:/home/ghost/ghost/slack-notifications/node_modules:delegated
-      - node_modules_ghost_stripe:/home/ghost/ghost/stripe/node_modules:delegated
       - node_modules_ghost_tiers:/home/ghost/ghost/tiers/node_modules:delegated
       - node_modules_ghost_webmentions:/home/ghost/ghost/webmentions/node_modules:delegated
       - node_modules_apps_admin-x-activitypub:/home/ghost/apps/admin-x-activitypub/node_modules:delegated
@@ -234,7 +233,6 @@ volumes:
   node_modules_ghost_recommendations: {}
   node_modules_ghost_security: {}
   node_modules_ghost_slack-notifications: {}
-  node_modules_ghost_stripe: {}
   node_modules_ghost_tiers: {}
   node_modules_ghost_webmentions: {}
   node_modules_apps_admin-x-activitypub: {}
