+++ ghost/core/.c8rc.e2e.json
@@ -7,10 +7,10 @@
     "cobertura"
   ],
   "reportsDir": "./coverage-e2e",
-  "statements": 80,
+  "statements": 79,
   "branches": 82,
   "functions": 81,
-  "lines": 80,
+  "lines": 79,
   "include": [
     "core/{*.js,frontend,server,shared}"
   ],
