+++ ghost/core/package.json
@@ -118,7 +118,6 @@
     "@tryghost/members-api": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
-    "@tryghost/members-stripe-service": "0.0.0",
     "@tryghost/metrics": "1.0.37",
     "@tryghost/milestones": "0.0.0",
     "@tryghost/minifier": "0.0.0",
@@ -193,6 +192,7 @@
     "keypair": "1.0.4",
     "knex": "2.4.2",
     "knex-migrator": "5.3.0",
+    "leaky-bucket": "2.2.0",
     "lodash": "4.17.21",
     "luxon": "3.6.0",
     "metascraper": "5.45.15",
@@ -221,6 +221,7 @@
     "semver": "7.7.1",
     "simple-dom": "1.4.0",
     "stoppable": "1.1.0",
+    "stripe": "8.222.0",
     "superagent": "5.1.0",
     "superagent-throttle": "1.0.1",
     "tough-cookie": "4.1.4",
