+++ ghost/core/test/unit/server/services/stripe/StripeAPI.test.js
@@ -1,7 +1,7 @@
 const sinon = require('sinon');
 const should = require('should');
 const rewire = require('rewire');
-const StripeAPI = rewire('../../../lib/StripeAPI');
+const StripeAPI = rewire('../../../../../core/server/services/stripe/StripeAPI');
 
 describe('StripeAPI', function () {
     const mockCustomerEmail = 'foo@example.com';
