+++ ghost/core/test/unit/server/services/stripe/services/webhooks/CheckoutSessionEventService.test.js
@@ -2,7 +2,7 @@ const assert = require('assert/strict');
 const errors = require('@tryghost/errors');
 const sinon = require('sinon');
 
-const CheckoutSessionEventService = require('../../../../../lib/services/webhook/CheckoutSessionEventService');
+const CheckoutSessionEventService = require('../../../../../../../core/server/services/stripe/services/webhook/CheckoutSessionEventService');
 
 describe('CheckoutSessionEventService', function () {
     let api, memberRepository, donationRepository, staffServiceEmails, sendSignupEmail;
