--- ghost/stripe/index.js
@@ -1,2 +1,0 @@
-module.exports = require('./lib/StripeService');
-module.exports.events = require('./lib/events');
