--- ghost/stripe/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@tryghost/members-stripe-service",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.2.0",
-    "rewire": "6.0.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/debug": "0.1.32",
-    "@tryghost/domain-events": "0.0.0",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/member-events": "0.0.0",
-    "leaky-bucket": "2.2.0",
-    "lodash": "4.17.21",
-    "stripe": "8.222.0"
-  }
-}
