--- ghost/stripe/test/hello.test.js
@@ -1,10 +1,0 @@
-// Switch these lines once there are useful utils
-// const testUtils = require('./utils');
-require('should');
-
-describe('Hello world', function () {
-    it('Runs a test', function () {
-        // TODO: Write me!
-        'hello'.should.eql('hello');
-    });
-});
